AccountId and new state, changes to JournalInner
