+++ crates/context/src/journal/inner.rs
@@ -12,9 +12,10 @@ use database_interface::Database;
 use primitives::{
     hardfork::SpecId::{self, *},
     hash_map::Entry,
-    Address, HashMap, Log, StorageKey, StorageValue, B256, KECCAK_EMPTY, U256,
+    AccountId, Address, AddressAndId, AddressOrId, HashMap, Log, StorageKey, StorageValue, B256,
+    KECCAK_EMPTY, U256,
 };
-use state::{Account, EvmState, EvmStorageSlot, TransientStorage};
+use state::{Account, EvmState, EvmStateNew, EvmStorageSlot, TransientStorage};
 use std::vec::Vec;
 /// Inner journal state that contains journal and state changes.
 ///
@@ -24,6 +25,8 @@ use std::vec::Vec;
 pub struct JournalInner<ENTRY> {
     /// The current state
     pub state: EvmState,
+    /// The current state new
+    pub state_new: EvmStateNew,
     /// Transient storage that is discarded after every transaction.
     ///
     /// See [EIP-1153](https://eips.ethereum.org/EIPS/eip-1153).
@@ -71,6 +74,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn new() -> JournalInner<ENTRY> {
         Self {
             state: HashMap::default(),
+            state_new: EvmStateNew::new(),
             transient_storage: TransientStorage::default(),
             logs: Vec::new(),
             journal: Vec::default(),
@@ -99,6 +103,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         // missing any field.
         let Self {
             state,
+            state_new,
             transient_storage,
             logs,
             depth,
@@ -128,6 +133,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         // if there is no journal entries, there has not been any changes.
         let Self {
             state,
+            state_new,
             transient_storage,
             logs,
             depth,
@@ -160,6 +166,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         // missing any field.
         let Self {
             state,
+            state_new,
             transient_storage,
             logs,
             depth,
@@ -202,17 +209,17 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     /// This is especially important for state clear where touched empty accounts needs to
     /// be removed from state.
     #[inline]
-    pub fn touch(&mut self, address: Address) {
-        if let Some(account) = self.state.get_mut(&address) {
-            Self::touch_account(&mut self.journal, address, account);
+    pub fn touch(&mut self, address_or_id: &AddressOrId) {
+        if let Some((account, id)) = self.state_new.get_mut(address_or_id) {
+            Self::touch_account(&mut self.journal, id.id(), account);
         }
     }
 
     /// Mark account as touched.
     #[inline]
-    fn touch_account(journal: &mut Vec<ENTRY>, address: Address, account: &mut Account) {
+    fn touch_account(journal: &mut Vec<ENTRY>, id: AccountId, account: &mut Account) {
         if !account.is_touched() {
-            journal.push(ENTRY::account_touched(address));
+            journal.push(ENTRY::account_touched(id));
             account.mark_touch();
         }
     }
@@ -225,21 +232,21 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     ///
     /// Panics if the account has not been loaded and is missing from the state set.
     #[inline]
-    pub fn account(&self, address: Address) -> &Account {
-        self.state
-            .get(&address)
-            .expect("Account expected to be loaded") // Always assume that acc is already loaded
+    pub fn account(&self, address_or_id: &AddressOrId) -> (&Account, AddressAndId) {
+        self.state_new
+            .get(&address_or_id)
+            .expect("Account expected to be loaded") // Always assume that acc is already loaded.
     }
 
     /// Set code and its hash to the account.
     ///
     /// Note: Assume account is warm and that hash is calculated from code.
     #[inline]
-    pub fn set_code_with_hash(&mut self, address: Address, code: Bytecode, hash: B256) {
-        let account = self.state.get_mut(&address).unwrap();
-        Self::touch_account(&mut self.journal, address, account);
+    pub fn set_code_with_hash(&mut self, address_or_id: AddressOrId, code: Bytecode, hash: B256) {
+        let (account, address) = self.state_new.get_mut(&address_or_id).unwrap();
+        Self::touch_account(&mut self.journal, address.id(), account);
 
-        self.journal.push(ENTRY::code_changed(address));
+        self.journal.push(ENTRY::code_changed(address.id()));
 
         account.info.code_hash = hash;
         account.info.code = Some(code);
@@ -251,35 +258,35 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     ///
     /// In case of EIP-7702 code with zero address, the bytecode will be erased.
     #[inline]
-    pub fn set_code(&mut self, address: Address, code: Bytecode) {
+    pub fn set_code(&mut self, address_or_id: AddressOrId, code: Bytecode) {
         if let Bytecode::Eip7702(eip7702_bytecode) = &code {
             if eip7702_bytecode.address().is_zero() {
-                self.set_code_with_hash(address, Bytecode::default(), KECCAK_EMPTY);
+                self.set_code_with_hash(address_or_id, Bytecode::default(), KECCAK_EMPTY);
                 return;
             }
         }
 
         let hash = code.hash_slow();
-        self.set_code_with_hash(address, code, hash)
+        self.set_code_with_hash(address_or_id, code, hash)
     }
 
     /// Add journal entry for caller accounting.
     #[inline]
     pub fn caller_accounting_journal_entry(
         &mut self,
-        address: Address,
+        address_or_id: AddressOrId,
         old_balance: U256,
         bump_nonce: bool,
     ) {
+        let id = self.state_new.get_id(&address_or_id).unwrap();
         // account balance changed.
-        self.journal
-            .push(ENTRY::balance_changed(address, old_balance));
+        self.journal.push(ENTRY::balance_changed(id, old_balance));
         // account is touched.
-        self.journal.push(ENTRY::account_touched(address));
+        self.journal.push(ENTRY::account_touched(id));
 
         if bump_nonce {
             // nonce changed.
-            self.journal.push(ENTRY::nonce_changed(address));
+            self.journal.push(ENTRY::nonce_changed(id));
         }
     }
 
@@ -290,29 +297,31 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn balance_incr<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
+        address_or_id: AddressOrId,
         balance: U256,
     ) -> Result<(), DB::Error> {
-        let account = self.load_account(db, address)?.data;
+        let (account, address) = self.load_account(db, address_or_id)?.data;
         let old_balance = account.info.balance;
         account.info.balance = account.info.balance.saturating_add(balance);
 
         // march account as touched.
         if !account.is_touched() {
             account.mark_touch();
-            self.journal.push(ENTRY::account_touched(address));
+            self.journal.push(ENTRY::account_touched(address.id()));
         }
 
         // add journal entry for balance increment.
         self.journal
-            .push(ENTRY::balance_changed(address, old_balance));
+            .push(ENTRY::balance_changed(address.id(), old_balance));
         Ok(())
     }
 
     /// Increments the nonce of the account.
     #[inline]
-    pub fn nonce_bump_journal_entry(&mut self, address: Address) {
-        self.journal.push(ENTRY::nonce_changed(address));
+    pub fn nonce_bump_journal_entry(&mut self, address_or_id: AddressOrId) {
+        // TODO check if it is okay to unwrap here
+        let id = self.state_new.get_id(&address_or_id).unwrap();
+        self.journal.push(ENTRY::nonce_changed(id));
     }
 
     /// Transfers balance from two accounts. Returns error if sender balance is not enough.
@@ -320,23 +329,23 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn transfer<DB: Database>(
         &mut self,
         db: &mut DB,
-        from: Address,
-        to: Address,
+        from: AddressOrId,
+        to: AddressOrId,
         balance: U256,
     ) -> Result<Option<TransferError>, DB::Error> {
         if balance.is_zero() {
             self.load_account(db, to)?;
-            let to_account = self.state.get_mut(&to).unwrap();
-            Self::touch_account(&mut self.journal, to, to_account);
+            let (to_account, to) = self.state_new.get_mut(&to).unwrap();
+            Self::touch_account(&mut self.journal, to.id(), to_account);
             return Ok(None);
         }
         // load accounts
-        self.load_account(db, from)?;
-        self.load_account(db, to)?;
+        let (_, to) = self.load_account(db, to)?.data;
+        let (_, from) = self.load_account(db, from)?.data;
 
         // sub balance from
-        let from_account = self.state.get_mut(&from).unwrap();
-        Self::touch_account(&mut self.journal, from, from_account);
+        let (from_account, from) = self.state_new.get_mut(&AddressOrId::Id(from.id())).unwrap();
+        Self::touch_account(&mut self.journal, from.id(), from_account);
         let from_balance = &mut from_account.info.balance;
 
         let Some(from_balance_decr) = from_balance.checked_sub(balance) else {
@@ -345,8 +354,8 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         *from_balance = from_balance_decr;
 
         // add balance to
-        let to_account = &mut self.state.get_mut(&to).unwrap();
-        Self::touch_account(&mut self.journal, to, to_account);
+        let (to_account, to) = self.state_new.get_mut(&AddressOrId::Id(to.id())).unwrap();
+        Self::touch_account(&mut self.journal, to.id(), to_account);
         let to_balance = &mut to_account.info.balance;
         let Some(to_balance_incr) = to_balance.checked_add(balance) else {
             return Ok(Some(TransferError::OverflowPayment));
@@ -355,7 +364,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         // Overflow of U256 balance is not possible to happen on mainnet. We don't bother to return funds from from_acc.
 
         self.journal
-            .push(ENTRY::balance_transfer(from, to, balance));
+            .push(ENTRY::balance_transfer(from.id(), to.id(), balance));
 
         Ok(None)
     }
@@ -378,7 +387,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     #[inline]
     pub fn create_account_checkpoint(
         &mut self,
-        caller: Address,
+        caller: AddressOrId,
         target_address: Address,
         balance: U256,
         spec_id: SpecId,
@@ -387,7 +396,8 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         let checkpoint = self.checkpoint();
 
         // Fetch balance of caller.
-        let caller_balance = self.state.get(&caller).unwrap().info.balance;
+        let (caller_acc, caller) = self.state_new.get(&caller).unwrap();
+        let caller_balance = caller_acc.info.balance;
         // Check if caller has enough balance to send to the created contract.
         if caller_balance < balance {
             self.checkpoint_revert(checkpoint);
@@ -395,7 +405,10 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         }
 
         // Newly created account is present, as we just loaded it.
-        let target_acc = self.state.get_mut(&target_address).unwrap();
+        let (target_acc, target) = self
+            .state_new
+            .get_mut(&AddressOrId::Address(target_address))
+            .unwrap();
         let last_journal = &mut self.journal;
 
         // New account can be created if:
@@ -411,7 +424,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         let is_created_globally = target_acc.mark_created_locally();
 
         // this entry will revert set nonce.
-        last_journal.push(ENTRY::account_created(target_address, is_created_globally));
+        last_journal.push(ENTRY::account_created(target.id(), is_created_globally));
         target_acc.info.code = None;
         // EIP-161: State trie clearing (invariant-preserving alternative)
         if spec_id.is_enabled_in(SPURIOUS_DRAGON) {
@@ -421,7 +434,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
 
         // touch account. This is important as for pre SpuriousDragon account could be
         // saved even empty.
-        Self::touch_account(last_journal, target_address, target_acc);
+        Self::touch_account(last_journal, target.id(), target_acc);
 
         // Add balance to created account, as we already have target here.
         let Some(new_balance) = target_acc.info.balance.checked_add(balance) else {
@@ -431,10 +444,15 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         target_acc.info.balance = new_balance;
 
         // safe to decrement for the caller as balance check is already done.
-        self.state.get_mut(&caller).unwrap().info.balance -= balance;
+        self.state_new
+            .get_mut(&AddressOrId::Id(caller.id()))
+            .unwrap()
+            .0
+            .info
+            .balance -= balance;
 
         // add journal entry of transferred balance
-        last_journal.push(ENTRY::balance_transfer(caller, target_address, balance));
+        last_journal.push(ENTRY::balance_transfer(caller.id(), target.id(), balance));
 
         Ok(checkpoint)
     }
@@ -489,25 +507,27 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn selfdestruct<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
+        address_or_id: AddressOrId,
         target: Address,
     ) -> Result<StateLoad<SelfDestructResult>, DB::Error> {
+        // TODO re check this logic
         let spec = self.spec;
-        let account_load = self.load_account(db, target)?;
+        let account_load = self.load_account(db, AddressOrId::Address(target))?;
         let is_cold = account_load.is_cold;
-        let is_empty = account_load.state_clear_aware_is_empty(spec);
+        let is_empty = account_load.0.state_clear_aware_is_empty(spec);
+        let target = account_load.1;
 
-        if address != target {
+        let (acc, address) = self.state_new.get_mut(&address_or_id).unwrap();
+        if address.address() != target.address() {
             // Both accounts are loaded before this point, `address` as we execute its contract.
             // and `target` at the beginning of the function.
-            let acc_balance = self.state.get(&address).unwrap().info.balance;
+            let acc_balance = acc.info.balance;
 
-            let target_account = self.state.get_mut(&target).unwrap();
-            Self::touch_account(&mut self.journal, target, target_account);
+            let target_account = self.state.get_mut(target.address()).unwrap();
+            Self::touch_account(&mut self.journal, address.id(), target_account);
             target_account.info.balance += acc_balance;
         }
 
-        let acc = self.state.get_mut(&address).unwrap();
         let balance = acc.info.balance;
 
         let destroyed_status = if !acc.is_selfdestructed() {
@@ -525,14 +545,14 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
             acc.mark_selfdestructed_locally();
             acc.info.balance = U256::ZERO;
             Some(ENTRY::account_destroyed(
-                address,
-                target,
+                address.id(),
+                target.id(),
                 destroyed_status,
                 balance,
             ))
-        } else if address != target {
+        } else if address.address() != target.address() {
             acc.info.balance = U256::ZERO;
-            Some(ENTRY::balance_transfer(address, target, balance))
+            Some(ENTRY::balance_transfer(address.id(), target.id(), balance))
         } else {
             // State is not changed:
             // * if we are after Cancun upgrade and
@@ -561,9 +581,9 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn load_account<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
-    ) -> Result<StateLoad<&mut Account>, DB::Error> {
-        self.load_account_optional(db, address, false, [])
+        address_or_id: AddressOrId,
+    ) -> Result<StateLoad<(&mut Account, AddressAndId)>, DB::Error> {
+        self.load_account_optional(db, address_or_id, false, [])
     }
 
     /// Loads account into memory. If account is EIP-7702 type it will additionally
@@ -577,12 +597,12 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn load_account_delegated<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
+        address_or_id: AddressOrId,
     ) -> Result<StateLoad<AccountLoad>, DB::Error> {
         let spec = self.spec;
         let is_eip7702_enabled = spec.is_enabled_in(SpecId::PRAGUE);
-        let account = self.load_account_optional(db, address, is_eip7702_enabled, [])?;
-        let is_empty = account.state_clear_aware_is_empty(spec);
+        let account = self.load_account_optional(db, address_or_id, is_eip7702_enabled, [])?;
+        let is_empty = account.0.state_clear_aware_is_empty(spec);
 
         let mut account_load = StateLoad::new(
             AccountLoad {
@@ -593,9 +613,9 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         );
 
         // load delegate code if account is EIP-7702
-        if let Some(Bytecode::Eip7702(code)) = &account.info.code {
+        if let Some(Bytecode::Eip7702(code)) = &account.0.info.code {
             let address = code.address();
-            let delegate_account = self.load_account(db, address)?;
+            let delegate_account = self.load_account(db, AddressOrId::Address(address))?;
             account_load.data.is_delegate_account_cold = Some(delegate_account.is_cold);
         }
 
@@ -612,9 +632,9 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn load_code<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
-    ) -> Result<StateLoad<&mut Account>, DB::Error> {
-        self.load_account_optional(db, address, true, [])
+        address_or_id: AddressOrId,
+    ) -> Result<StateLoad<(&mut Account, AddressAndId)>, DB::Error> {
+        self.load_account_optional(db, address_or_id, true, [])
     }
 
     /// Loads account. If account is already loaded it will be marked as warm.
@@ -622,50 +642,53 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn load_account_optional<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
+        address_or_id: AddressOrId,
         load_code: bool,
         storage_keys: impl IntoIterator<Item = StorageKey>,
-    ) -> Result<StateLoad<&mut Account>, DB::Error> {
-        let load = match self.state.entry(address) {
-            Entry::Occupied(entry) => {
-                let account = entry.into_mut();
-                let is_cold = account.mark_warm_with_transaction_id(self.transaction_id);
-                // if it is colad loaded we need to clear local flags that can interact with selfdestruct
-                if is_cold {
-                    // if it is cold loaded and we have selfdestructed locally it means that
-                    // account was selfdestructed in previous transaction and we need to clear its information and storage.
-                    if account.is_selfdestructed_locally() {
-                        account.selfdestruct();
-                        account.unmark_selfdestructed_locally();
-                    }
-                    // unmark locally created
-                    account.unmark_created_locally();
-                }
-                StateLoad {
-                    data: account,
-                    is_cold,
-                }
-            }
-            Entry::Vacant(vac) => {
-                let account = if let Some(account) = db.basic(address)? {
-                    account.into()
-                } else {
-                    Account::new_not_existing(self.transaction_id)
-                };
+    ) -> Result<StateLoad<(&mut Account, AddressAndId)>, DB::Error> {
+        todo!();
+        let load = Default::default();
+        let id = AddressAndId::new(address_or_id.address().unwrap(), 0);
+        // let load = match self.state.entry(address) {
+        //     Entry::Occupied(entry) => {
+        //         let account = entry.into_mut();
+        //         let is_cold = account.mark_warm_with_transaction_id(self.transaction_id);
+        //         // if it is colad loaded we need to clear local flags that can interact with selfdestruct
+        //         if is_cold {
+        //             // if it is cold loaded and we have selfdestructed locally it means that
+        //             // account was selfdestructed in previous transaction and we need to clear its information and storage.
+        //             if account.is_selfdestructed_locally() {
+        //                 account.selfdestruct();
+        //                 account.unmark_selfdestructed_locally();
+        //             }
+        //             // unmark locally created
+        //             account.unmark_created_locally();
+        //         }
+        //         StateLoad {
+        //             data: account,
+        //             is_cold,
+        //         }
+        //     }
+        //     Entry::Vacant(vac) => {
+        //         let account = if let Some(account) = db.basic(address)? {
+        //             account.into()
+        //         } else {
+        //             Account::new_not_existing(self.transaction_id)
+        //         };
 
-                // Precompiles among some other account(coinbase included) are warm loaded so we need to take that into account
-                let is_cold = self.warm_addresses.is_cold(&address);
+        //         // Precompiles among some other account(coinbase included) are warm loaded so we need to take that into account
+        //         let is_cold = self.warm_addresses.is_cold(&address);
 
-                StateLoad {
-                    data: vac.insert(account),
-                    is_cold,
-                }
-            }
-        };
+        //         StateLoad {
+        //             data: vac.insert(account),
+        //             is_cold,
+        //         }
+        //     }
+        // };
 
         // journal loading of cold account.
         if load.is_cold {
-            self.journal.push(ENTRY::account_warmed(address));
+            self.journal.push(ENTRY::account_warmed(id.id()));
         }
         if load_code {
             let info = &mut load.data.info;
@@ -685,11 +708,11 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
                 db,
                 &mut self.journal,
                 self.transaction_id,
-                address,
+                id,
                 storage_key,
             )?;
         }
-        Ok(load)
+        Ok((load, id))
     }
 
     /// Loads storage slot.
@@ -701,20 +724,13 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn sload<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
+        address_or_id: AddressOrId,
         key: StorageKey,
     ) -> Result<StateLoad<StorageValue>, DB::Error> {
         // assume acc is warm
-        let account = self.state.get_mut(&address).unwrap();
+        let (account, id) = self.state_new.get_mut(&address_or_id).unwrap();
         // only if account is created in this tx we can assume that storage is empty.
-        sload_with_account(
-            account,
-            db,
-            &mut self.journal,
-            self.transaction_id,
-            address,
-            key,
-        )
+        sload_with_account(account, db, &mut self.journal, self.transaction_id, id, key)
     }
 
     /// Stores storage slot.
@@ -726,13 +742,13 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     pub fn sstore<DB: Database>(
         &mut self,
         db: &mut DB,
-        address: Address,
+        address_or_id: AddressOrId,
         key: StorageKey,
         new: StorageValue,
     ) -> Result<StateLoad<SStoreResult>, DB::Error> {
         // assume that acc exists and load the slot.
-        let present = self.sload(db, address, key)?;
-        let acc = self.state.get_mut(&address).unwrap();
+        let present = self.sload(db, address_or_id, key)?;
+        let (acc, id) = self.state_new.get_mut(&address_or_id).unwrap();
 
         // if there is no original value in dirty return present value, that is our original.
         let slot = acc.storage.get_mut(&key).unwrap();
@@ -750,7 +766,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         }
 
         self.journal
-            .push(ENTRY::storage_changed(address, key, present.data));
+            .push(ENTRY::storage_changed(id.id(), key, present.data));
         // insert value into present state.
         slot.present_value = new;
         Ok(StateLoad::new(
@@ -767,9 +783,10 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     ///
     /// EIP-1153: Transient storage opcodes
     #[inline]
-    pub fn tload(&mut self, address: Address, key: StorageKey) -> StorageValue {
+    pub fn tload(&mut self, address_or_id: AddressOrId, key: StorageKey) -> StorageValue {
+        let id = self.state_new.get_id(&address_or_id).unwrap();
         self.transient_storage
-            .get(&(address, key))
+            .get(&(id, key))
             .copied()
             .unwrap_or_default()
     }
@@ -781,17 +798,18 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
     ///
     /// EIP-1153: Transient storage opcodes
     #[inline]
-    pub fn tstore(&mut self, address: Address, key: StorageKey, new: StorageValue) {
+    pub fn tstore(&mut self, address_or_id: AddressOrId, key: StorageKey, new: StorageValue) {
+        let id = self.state_new.get_id(&address_or_id).unwrap();
         let had_value = if new.is_zero() {
             // if new values is zero, remove entry from transient storage.
             // if previous values was some insert it inside journal.
             // If it is none nothing should be inserted.
-            self.transient_storage.remove(&(address, key))
+            self.transient_storage.remove(&(id, key))
         } else {
             // insert values
             let previous_value = self
                 .transient_storage
-                .insert((address, key), new)
+                .insert((id, key), new)
                 .unwrap_or_default();
 
             // check if previous value is same
@@ -806,7 +824,7 @@ impl<ENTRY: JournalEntryTr> JournalInner<ENTRY> {
         if let Some(had_value) = had_value {
             // insert in journal only if value was changed.
             self.journal
-                .push(ENTRY::transient_storage_changed(address, key, had_value));
+                .push(ENTRY::transient_storage_changed(id, key, had_value));
         }
     }
 
@@ -824,7 +842,7 @@ pub fn sload_with_account<DB: Database, ENTRY: JournalEntryTr>(
     db: &mut DB,
     journal: &mut Vec<ENTRY>,
     transaction_id: usize,
-    address: Address,
+    address_or_id: AddressAndId,
     key: StorageKey,
 ) -> Result<StateLoad<StorageValue>, DB::Error> {
     let is_newly_created = account.is_created();
@@ -839,7 +857,7 @@ pub fn sload_with_account<DB: Database, ENTRY: JournalEntryTr>(
             let value = if is_newly_created {
                 StorageValue::ZERO
             } else {
-                db.storage(address, key)?
+                db.storage(*address_or_id.address(), key)?
             };
 
             vac.insert(EvmStorageSlot::new(value, transaction_id));
@@ -850,7 +868,7 @@ pub fn sload_with_account<DB: Database, ENTRY: JournalEntryTr>(
 
     if is_cold {
         // add it to journal as cold loaded.
-        journal.push(ENTRY::storage_warmed(address, key));
+        journal.push(ENTRY::storage_warmed(address_or_id.id(), key));
     }
 
     Ok(StateLoad::new(value, is_cold))
